+++ kinto_attachment/utils.py
@@ -217,3 +217,16 @@ def save_file(content, request, randomize=True, gzipped=False):
     })
 
     return attachment
+
+
+def setting_value(request, name, default):
+    value = request.registry.settings.get('attachment.{}'.format(name), default)
+    if 'bucket_id' in request.matchdict:
+        uri = '/buckets/{bucket_id}'.format(**request.matchdict)
+        if uri in request.registry.attachment_resources:
+            value = request.registry.attachment_resources[uri].get(name, value)
+        if 'collection_id' in request.matchdict:
+            uri = '/buckets/{bucket_id}/collections/{collection_id}'.format(**request.matchdict)
+            if uri in request.registry.attachment_resources:
+                value = request.registry.attachment_resources[uri].get(name, value)
+    return value