+++ CHANGELOG.rst
@@ -9,6 +9,9 @@ Changelog
 - Gzip ``Content-Encoding`` is now always enabled when uploading on S3 (fixes #139)
 - Overriding settings via the querystring (eg. ``?gzipped``, ``randomize``, ``use_content_encoding``) is not possible anymore
 
+**Internal changes**
+
+- Refactor reading of settings
 
 3.0.1 (2018-07-05)
 ------------------